+++ packages/ibm-products/src/components/EditInPlace/EditInPlace.stories.js
@@ -11,20 +11,59 @@ import {
   prepareStory,
 } from '../../global/js/utils/story-helper';
 import { action } from '@storybook/addon-actions';
-import { InlineEdit } from '../InlineEdit/InlineEdit';
-import { InlineEditV2 } from '.';
-import mdx from './InlineEditV2.mdx';
+import { EditInPlace } from '.';
+import { DisplayBox } from '../../global/js/utils/DisplayBox';
+import mdx from './EditInPlace.mdx';
 import styles from './_storybook-styles.scss';
 
+const storyClass = 'edit-in-place-example';
+
+const tooltipAlignmentOptions = {
+  'Default / undefined': undefined,
+  'All top': 'top',
+  'All top-left': 'top-left',
+  'All top-right': 'top-right',
+  'All bottom': 'bottom',
+  'All bottom-left': 'bottom-left',
+  'All bottom-right': 'bottom-right',
+  'All left': 'left',
+  'All right': 'right',
+  'Edit and save right, cancel left': {
+    edit: 'right',
+    cancel: 'left',
+    save: 'right',
+  },
+};
+
 export default {
-  title: getStoryTitle(InlineEditV2.displayName),
-  component: InlineEditV2,
+  title: getStoryTitle(EditInPlace.displayName),
+  component: EditInPlace,
+  argTypes: {
+    containerWidth: {
+      control: { type: 'range', min: 20, max: 800, step: 10 },
+      description:
+        'Controls containing element width. Used for demonstration purposes, not property of the component.',
+    },
+    tooltipAlignment: {
+      control: {
+        type: 'select',
+        labels: Object.keys(tooltipAlignmentOptions),
+      },
+      options: Object.values(tooltipAlignmentOptions).map((_k, i) => i),
+      mapping: Object.values(tooltipAlignmentOptions),
+    },
+  },
   parameters: {
     styles,
     docs: {
       page: mdx,
     },
   },
+  decorators: [
+    (story) => (
+      <DisplayBox className={`${storyClass}__viewport`}>{story()}</DisplayBox>
+    ),
+  ],
 };
 
 const actionSave = action('save');
@@ -33,6 +72,7 @@ const actionCancel = action('cancel');
 
 const defaultProps = {
   cancelLabel: 'Cancel',
+  containerWidth: 300,
   editLabel: 'Edit',
   id: 'story-id',
   invalid: false,
@@ -47,7 +87,7 @@ const defaultProps = {
   value: 'default',
 };
 
-const Template = (args) => {
+const Template = ({ containerWidth, ...args }) => {
   const [value, setValue] = useState(defaultProps.value);
 
   const onChange = (val) => {
@@ -73,7 +113,11 @@ const Template = (args) => {
     onCancel,
   };
 
-  return <InlineEdit {...props} className="inline-edit-v2-example" />;
+  return (
+    <div style={{ width: containerWidth }}>
+      <EditInPlace {...props} className="edit-in-place-example" />
+    </div>
+  );
 };
 
 export const Default = prepareStory(Template, {