+++ packages/ibm-products/src/components/PageHeader/PageHeader.js
@@ -1082,12 +1082,14 @@ PageHeader.propTypes = {
       loading: PropTypes.bool,
 
       // inline edit version properties
-      editableLabel: PropTypes.string, // .isRequired.if(inlineEditRequired),
-      id: PropTypes.string, // .isRequired.if(inlineEditRequired),
+      editableLabel: PropTypes.string, // .isRequired.if(editInPlaceRequired),
+      id: PropTypes.string, // .isRequired.if(editInPlaceRequired),
+      onCancel: PropTypes.func,
       onChange: PropTypes.func,
       onSave: PropTypes.func,
-      cancelDescription: PropTypes.string, //.isRequired.if(inlineEditRequired),
-      saveDescription: PropTypes.string, //.isRequired.if(inlineEditRequired),
+      cancelDescription: PropTypes.string, //.isRequired.if(editInPlaceRequired),
+      editDescription: PropTypes.string, // .isRequired.if(editInPlaceRequired),
+      saveDescription: PropTypes.string, //.isRequired.if(editInPlaceRequired),
       // Update docgen if changed
     }),
     PropTypes.string,